Server: update XCTest/ headers from WebDriverAgent
* 143332260c53561bb7e544ae4e2d4b8dc649ac8d
* Tue Dec 20 9:53 CET

+++ Server/CBXConstants.h
@@ -112,4 +112,4 @@ static float const CBX_FINGER_WIDTH = 78;
 static int const CBX_DEFAULT_NUM_FINGERS = 1;
 static int const CBX_DEFAULT_REPETITIONS = 1;
 
-static unsigned long long CBX_DEFAULT_SEND_STRING_FREQUENCY = 60; //FB uses 60
+static NSUInteger CBX_DEFAULT_SEND_STRING_FREQUENCY = 60; //FB uses 60
